// src/interfaces/stripe.ts
export interface ISaleInfo {
  email: string;
  firstNameOnCard: string;
  lastNameOnCard: string;
  card: string; // card number
  ccExpMonth: string;
  ccExpYear: string;
  chargeAmount: number;
  cvc: string;
  uuid: string;
}

export class SaleInfo implements ISaleInfo {
  constructor(
    public email: string,
    public firstNameOnCard: string,
    public lastNameOnCard: string,
    public card: string,
    public ccExpMonth: string,
    public ccExpYear: string,
    public cvc: string,
    public chargeAmount: number,
    public uuid: string
  ) {}
}

export interface IChargeCustomerResult {
  success: boolean;
  customerId: string;
  sourceId: string;
  chargeInfo: Stripe.charges.ICharge;
}

export interface IStripeChargeCustomerResult {
  success: boolean;
  payload?: IChargeCustomerResult;
  exception?: IException;
}